+++ templates/finance/balances/balances.html
@@ -13,7 +13,7 @@
         <div class="col-12" style="height: 50px;">
             <a class="btn btn-primary" href="{% url 'finance:new_balance' %}" role="button">Novo</a>
             <button type="button" class="btn btn-danger" onclick="deleteBalances(event, this)">Excluir</button>
-            <a class="btn btn-success" href="{% url 'finance:sync_balances' %}" role="button">Sincronizar</a>
+            <!-- <a class="btn btn-success" href="{% url 'finance:sync_balances' %}" role="button">Sincronizar</a> -->
         </div>
     </div>
     {% if messages %}
