+++ templates/finance/variable_expenses/variable_expenses.html
@@ -13,20 +13,41 @@
         <div class="col-12" style="height: 50px;">
             <a class="btn btn-primary" href="{% url 'finance:new_variable_expense' %}" role="button">Novo</a>
             <button type="button" class="btn btn-danger" onclick="deleteExpenses(event, this)">Excluir</button>
-            <a class="btn btn-success" href="{% url 'finance:sync_variable_expenses' %}" role="button">Sincronizar</a>
+            <!-- <a class="btn btn-success" href="{% url 'finance:sync_variable_expenses' %}" role="button">Sincronizar</a> -->
+            <form id="nubank-import-form" action="{% url 'finance:import_variable_expenses_nubank' %}" method="post" enctype="multipart/form-data" style="display:inline;">
+                {% csrf_token %}
+                <input type="file" id="nubank-csv-input" name="csv_file" accept=".csv" style="display:none;" onchange="document.getElementById('nubank-import-form').submit();">
+                <button type="button" class="btn" style="background-color: #820ad1; color: #fff; display: inline-flex; align-items: center; border: none;" onclick="document.getElementById('nubank-csv-input').click();">
+                    <img src="{% static 'img/nubank_logo.png' %}" alt="Nubank" style="height: 24px; margin-right: 8px;"> Importar Nubank
+                </button>
+            </form>
         </div>
     </div>
-    {% if messages %}
-        <div id="error-alert" class="alert alert-danger" role="alert">
+    {% if messages.error %}
+        <div id="alert-error" class="alert alert-danger" role="alert">
             <ul class="mb-0">
-                {% for message in messages %}
+                {% for message in messages.error %}
                     <li>{{ message }}</li>
                 {% endfor %}
             </ul>
         </div>
         <script>
             setTimeout(function() {
-                var alert = document.getElementById('error-alert');
+                var alert = document.getElementById('alert-error');
+                if (alert) { alert.style.display = 'none'; }
+            }, 5000);
+        </script>
+    {% elif messages.success %}
+        <div id="alert-success" class="alert alert-success" role="alert">
+            <ul class="mb-0">
+                {% for message in messages.success %}
+                    <li>{{ message }}</li>
+                {% endfor %}
+            </ul>
+        </div>
+        <script>
+            setTimeout(function() {
+                var alert = document.getElementById('alert-success');
                 if (alert) { alert.style.display = 'none'; }
             }, 5000);
         </script>
